разбил код на куски
